+++ src/RevisionPopupWidget.js
@@ -27,7 +27,7 @@ OO.inheritClass( RevisionPopupWidget, OO.ui.PopupWidget );
  * @param {number} size
  * @return {jQuery}
  */
-function getSizeHtml( size ) {
+function getSizeContent( size ) {
 	const $diffBytes = $( '<span>' );
 	$diffBytes.addClass( 'mw-diff-bytes' );
 
@@ -54,31 +54,18 @@ function getSizeHtml( size ) {
  * returned HTML will start with a <br>; if there is no summary, it wont (so it can be appended to
  * the revision author information).
  * @param {Object} data As returned by Api.prototype.getTokenInfo().
- * @param {boolean} [isCached] Comment promise is cached
  * @return {jQuery}
  */
-function getCommentHtml( data, isCached ) {
+function getCommentContent( data ) {
 	const $revCommentSpan = $( '<span>' )
 		.addClass( 'wwt-revisionPopupWidget-comment' )
-		.append( $( '<span>' ).text( ' ' ), getSizeHtml( data.size ) );
+		.append( $( '<span>' ).text( ' ' ), getSizeContent( data.size ) );
 
-	if ( data.comment === undefined && !isCached ) {
-		// Not yet available.
-		const $shimmerSpan = $( '<span>' )
-			.addClass( 'wwt-shimmer wwt-shimmer-animation' );
-		$revCommentSpan.append( $shimmerSpan, $shimmerSpan.clone() );
-	} else if ( data.comment !== '' ) {
+	if ( data.comment !== '' ) {
 		const $commentSpan = $( '<span>' )
 			.addClass( 'comment comment--without-parentheses' )
 			.append( Tools.bidiIsolate( $.parseHTML( data.comment ) ) );
 		$revCommentSpan.prepend( $( '<br>' ), $commentSpan );
-
-		if ( !isCached ) {
-			$commentSpan
-				.addClass( 'wwt-revisionPopupWidget-comment-animated' );
-			$revCommentSpan
-				.addClass( 'wwt-revisionPopupWidget-comment-transparent' );
-		}
 	}
 	return $revCommentSpan;
 }
@@ -88,7 +75,7 @@ function getCommentHtml( data, isCached ) {
  * @param {Object} data As returned by Api.prototype.getTokenInfo().
  * @return {jQuery}
  */
-function getUserLinksHtml( data ) {
+function getUserLinksContent( data ) {
 	const contribsUrl = mw.util.getUrl( `Special:Contributions/${data.username}` ),
 		// We typically link to Special:Contribs for IPs.
 		userPageUrl = data.isIP ? contribsUrl : mw.util.getUrl( `User:${data.username}` );
@@ -125,14 +112,23 @@ function getUserLinksHtml( data ) {
 }
 
 /**
- * Show the revision popup based on the given token data, above the given element.
- * Note that the English namespaces will normalize to the wiki's local namespaces.
- * @param {Object} data As returned by Api.prototype.getTokenInfo().
- * @param {jQuery} $target Element the popup should be attached to.
- * @param {boolean} [isCached] Whether the comment promise is available, cached.
+ * Get content for the loading state.
+ * @return {jQuery}
  */
-RevisionPopupWidget.prototype.show = function ( data, $target, isCached ) {
-	const $userLinks = getUserLinksHtml( data ),
+function getLoadingStateContent() {
+	const $shimmerSpan = $( '<span>' )
+		.addClass( 'wwt-shimmer wwt-shimmer-animation' );
+	return $( '<div>' )
+		.append( $shimmerSpan, $shimmerSpan.clone(), $shimmerSpan.clone() );
+}
+
+/**
+ * Get content for the popup, called once data has been retrieved from the API.
+ * @param {Object} data
+ * @return {jQuery}
+ */
+function getContent( data ) {
+	const $userLinks = getUserLinksContent( data ),
 		dateStr = moment( data.revisionTime ).locale( mw.config.get( 'wgUserLanguage' ) ).format( 'LLL' ),
 		// Use jQuery to make sure attributes are properly escaped
 		$diffLink = $( '<a>' )
@@ -146,11 +142,37 @@ RevisionPopupWidget.prototype.show = function ( data, $target, isCached ) {
 		$scorePercent = $( '<strong>' ).text( mw.msg( scoreMsgKey + '-percent', data.score ) ),
 		$attributionMsg = $( '<div>' )
 			.addClass( 'wwt-revisionPopupWidget-attribution' )
-			.html( Tools.i18nHtml( scoreMsgKey, $scorePercent ) ),
-		$html = $( '<div>' )
-			.append( addedMsg, getCommentHtml( data, isCached ), $attributionMsg );
-	this.$popupContent.empty().append( $html );
+			.html( Tools.i18nHtml( scoreMsgKey, $scorePercent ) );
 
+	return $( '<div>' )
+		.append( addedMsg, getCommentContent( data ), $attributionMsg );
+}
+
+/**
+ * Show the revision popup based on the given token data, above the given element.
+ * Note that the English namespaces will normalize to the wiki's local namespaces.
+ * @param {Object} data As returned by Api.prototype.getTokenInfo().
+ * @param {jQuery} $target Element the popup should be attached to.
+ * @param {boolean} [isCached] Whether the comment promise is available, cached.
+ */
+RevisionPopupWidget.prototype.show = function ( data, $target, isCached ) {
+	let $content;
+
+	// If data.comment is undefined, we are in the loading state.
+	if ( data.comment === undefined ) {
+		$content = getLoadingStateContent();
+	} else {
+		$content = getContent( data );
+
+		if ( !isCached ) {
+			$content.addClass( 'wwt-revisionPopupWidget-animate' );
+		}
+	}
+
+	this.$popupContent.empty().append( $content );
+
+	// Ensure the popup is attached to the visible content,
+	// which it otherwise wouldn't be for image thumbnails.
 	if ( $target.find( '.thumb' ).length ) {
 		$target = $target.find( '.thumb' );
 	}
@@ -162,10 +184,9 @@ RevisionPopupWidget.prototype.show = function ( data, $target, isCached ) {
 	this.setFloatableContainer( $target );
 	this.toggle( true );
 
-	// Animate edit summary, if present.
-	if ( data.comment ) {
-		$( '.wwt-revisionPopupWidget-comment' ).removeClass( 'wwt-revisionPopupWidget-comment-transparent' );
-	}
+	// Animate content once loaded.
+	$( '.wwt-revisionPopupWidget-animate' )
+		.addClass( 'wwt-revisionPopupWidget-animate-visible' );
 };
 
 export default RevisionPopupWidget;